+++ packages/dev/gui/src/2D/controls/container.ts
@@ -365,8 +365,8 @@ export class Container extends Control {
             if (this.shadowBlur || this.shadowOffsetX || this.shadowOffsetY) {
                 context.shadowColor = this.shadowColor;
                 context.shadowBlur = this.shadowBlur;
-                context.shadowOffsetX = this.shadowOffsetX;
-                context.shadowOffsetY = this.shadowOffsetY;
+                context.shadowOffsetX = this.shadowOffsetX * this._host.idealRatio;
+                context.shadowOffsetY = this.shadowOffsetY * this._host.idealRatio;
             }
 
             context.fillStyle = this._getBackgroundColor(context);