+++ packages/public/glTF2Interface/babylon.glTF2Interface.d.ts
@@ -1080,21 +1080,29 @@ declare module BABYLON.GLTF2 {
         };
     }
 
+    /**
+     * Interfaces from the KHR_materials_clearcoat extension
+     */
     /** @internal */
-    interface IKHRMaterialsClearcoatIor {
-        clearcoatIor?: number;
-    }
-
-    /** @internal */
-    interface IKHRMaterialsClearcoatDarkening {
-        clearcoatDarkeningFactor?: number;
-        clearcoatDarkeningTexture?: ITextureInfo;
-    }
-
-    /** @internal */
-    interface IKHRMaterialsClearcoatColor {
-        clearcoatColorFactor?: number[];
-        clearcoatColorTexture?: ITextureInfo;
+    interface IKHRMaterialsCoat {
+        coatFactor?: number;
+        coatTexture?: ITextureInfo;
+        coatRoughnessFactor?: number;
+        coatRoughnessTexture?: ITextureInfo;
+        coatNormalTexture?: IMaterialNormalTextureInfo;
+        coatIor?: number;
+        coatDarkeningFactor?: number;
+        coatColorFactor?: number[];
+        coatColorTexture?: ITextureInfo;
+        coatAnisotropyStrength?: number;
+        coatAnisotropyRotation?: number;
+        coatAnisotropyTexture?: ITextureInfo;
+        /**
+         * Dictionary object with extension-specific objects
+         */
+        extensions?: {
+            [key: string]: any;
+        };
     }
 
     /** @internal */
@@ -1231,6 +1239,20 @@ declare module BABYLON.GLTF2 {
         sheenColorTexture?: ITextureInfo;
         sheenRoughnessFactor?: number;
         sheenRoughnessTexture?: ITextureInfo;
+    }
+
+    /**
+     * Interfaces from the KHR_materials_fuzz extension
+     */
+
+    /** @internal */
+    interface IKHRMaterialsFuzz {
+        fuzzFactor?: number;
+        fuzzTexture?: ITextureInfo;
+        fuzzColorFactor?: number[];
+        fuzzColorTexture?: ITextureInfo;
+        fuzzRoughnessFactor?: number;
+        fuzzRoughnessTexture?: ITextureInfo;
     }
 
     /**